web: Move autoplay/unmuteOverlay from Config to BaseLoadOptions - and made config default the load options
