+++ web/packages/core/src/load-options.ts
@@ -1,3 +1,46 @@
+export enum AutoPlay {
+    /**
+     * The player should automatically play the movie as soon as it is loaded.
+     *
+     * If the browser does not support automatic audio, the movie will begin
+     * muted.
+     */
+    On = "on",
+
+    /**
+     * The player should not attempt to automatically play the movie.
+     *
+     * This will leave it to the user or API to actually play when appropriate.
+     */
+    Off = "off",
+
+    /**
+     * The player should automatically play the movie as soon as it is deemed
+     * "appropriate" to do so.
+     *
+     * The exact behaviour depends on the browser, but commonly requires some
+     * form of user interaction on the page in order to allow auto playing videos
+     * with sound.
+     */
+    Auto = "auto",
+}
+
+/**
+ * When the player is muted, this controls whether or not Ruffle will show a
+ * "click to unmute" overlay on top of the movie.
+ */
+export enum UnmuteOverlay {
+    /**
+     * Show an overlay explaining that the movie is muted.
+     */
+    Visible = "visible",
+
+    /**
+     * Don't show an overlay and pretend that everything is fine.
+     */
+    Hidden = "hidden",
+}
+
 /**
  * Any options used for loading a movie.
  */
@@ -11,6 +54,21 @@ export interface BaseLoadOptions {
      * by any explicitly set here.
      */
     parameters?: URLSearchParams | string | Record<string, string>;
+
+    /**
+     * Controls the auto-play behaviour of Ruffle.
+     *
+     * @default AutoPlay.Auto
+     */
+    autoplay?: AutoPlay;
+
+    /**
+     * Controls the visibility of the unmute overlay when the player
+     * is started muted.
+     *
+     * @default UnmuteOverlay.Visible
+     */
+    unmuteOverlay?: UnmuteOverlay;
 }
 
 /**
